extract cases
